+++ src/main/java/ml/pluto7073/bartending/TheArtOfBartending.java
@@ -9,9 +9,8 @@ import ml.pluto7073.bartending.content.gui.BartendingMenuTypes;
 import ml.pluto7073.bartending.content.item.BartendingItems;
 import ml.pluto7073.bartending.content.sound.BartendingSounds;
 import ml.pluto7073.bartending.foundations.BartendingStats;
-import ml.pluto7073.bartending.foundations.alcohol.AbsorbedAlcoholHandler;
-import ml.pluto7073.bartending.foundations.alcohol.BartendingEntityData;
 import ml.pluto7073.bartending.foundations.alcohol.AlcoholHandler;
+import ml.pluto7073.bartending.foundations.alcohol.BartendingEntityData;
 import ml.pluto7073.bartending.foundations.command.BartendingCommands;
 import ml.pluto7073.bartending.foundations.config.BartendingCommonConfig;
 import ml.pluto7073.bartending.foundations.item.BartendingCreativeTabs;
@@ -40,7 +39,6 @@ public class TheArtOfBartending implements ModInitializer {
         PDCommonConfig.INSTANCE.addManagedConfig(CONFIG_TYPE);
         GlassDrinkBaseSerializer.init();
         BartendingEntityData.init();
-        AbsorbedAlcoholHandler.init();
         AlcoholHandler.init();
         Arrays.stream(BartendingRecipes.values()).forEach(BartendingRecipes::register);
         Arrays.stream(BartendingStats.values()).forEach(BartendingStats::register);